# services/api/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "SentimentPoint": {
        "properties": {
          "avg_score": {
            "title": "Avg Score",
            "type": "number"
          },
          "bucket_start": {
            "format": "date-time",
            "title": "Bucket Start",
            "type": "string"
          },
          "n": {
            "title": "N",
            "type": "integer"
          }
        },
        "required": [
          "bucket_start",
          "n",
          "avg_score"
        ],
        "title": "SentimentPoint",
        "type": "object"
      },
      "SentimentSeries": {
        "properties": {
          "bucket": {
            "title": "Bucket",
            "type": "string"
          },
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "period": {
            "title": "Period",
            "type": "string"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/SentimentPoint"
            },
            "title": "Series",
            "type": "array"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "tool",
          "period",
          "bucket",
          "start",
          "end",
          "series"
        ],
        "title": "SentimentSeries",
        "type": "object"
      },
      "Tool": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "display_name",
          "aliases"
        ],
        "title": "Tool",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Vibecheck API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/sentiment/{tool}": {
      "get": {
        "operationId": "tool_sentiment_api_v1_sentiment__tool__get",
        "parameters": [
          {
            "in": "path",
            "name": "tool",
            "required": true,
            "schema": {
              "title": "Tool",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "30d",
              "title": "Period",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "bucket",
            "required": false,
            "schema": {
              "default": "day",
              "title": "Bucket",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SentimentSeries"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Tool not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tool Sentiment"
      }
    },
    "/api/v1/tools": {
      "get": {
        "operationId": "list_tools_api_v1_tools_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Tool"
                  },
                  "title": "Response List Tools Api V1 Tools Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Tools"
      }
    },
    "/health": {
      "get": {
        "description": "Liveness probe \u2014 does not touch the database.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}